fix : matrix_check,vector_filter and class_check moved to pycm_util
